% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWlasso-package.R
\docType{package}
\name{GWlasso-package}
\alias{GWlasso}
\alias{GWlasso-package}
\title{GWlasso: Geographically Weighted Lasso}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matthieu Mulot \email{matthieu.mulot@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Sophie Erb (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
