% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gwlfit.R
\name{plot.gwlfit}
\alias{plot.gwlfit}
\title{Plot method for gwlfit object}
\usage{
\method{plot}{gwlfit}(x, ...)
}
\arguments{
\item{x}{a gwlfit object returnes by \code{\link[=gwl_fit]{gwl_fit()}}}

\item{...}{ellipsis for S3 method compatibility}
}
\value{
a ggplot
}
\description{
Plot method for gwlfit object
}
