% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gwl_map.R
\name{plot_gwl_map}
\alias{plot_gwl_map}
\title{Plot a map of beta coefficient for gwlfit object}
\usage{
plot_gwl_map(x, column, crs = 4326)
}
\arguments{
\item{x}{a gwlfit object returned by \code{\link[=gwl_fit]{gwl_fit()}}.}

\item{column}{the name of a variable to be plotted on the map. Must be quotted. for instance "NEB.MIN"}

\item{crs}{the crs projection for the map (default is mercator WGS84)}
}
\value{
a ggplot object
}
\description{
this function plot a map of the beta coefficient for a selected column (aka species).
For this function to work, the coordinates supplied to \code{\link[=gwl_fit]{gwl_fit()}} must be named "Long" and "Lat".
The function is not bulletproof yet but is added here to reproduce the maps from the original publication.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{


if(interactive()){
  data(Amesbury)
  
  distance_matrix <- compute_distance_matrix(Amesbury$coords, add.noise = TRUE)
  
 
  my.gwl.fit <- gwl_fit(bw= 120,
                        x.var = Amesbury$spe.df,
                        y.var = Amesbury$WTD,
                        dist.mat = distance_matrix,
                        adaptive = TRUE,
                        kernel = "bisquare",
                        alpha = 1,
                        progress = TRUE)
  
  plot(my.gwl.fit)
}

}
