% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gwlfit.R
\name{predict.gwlfit}
\alias{predict.gwlfit}
\title{Predict method for GWL fits}
\usage{
\method{predict}{gwlfit}(object, newdata, newcoords, type = "response")
}
\arguments{
\item{object}{Object of class inheriting from "gwlfit"}

\item{newdata}{a data.frame or matrix with the same columns as the training dataset}

\item{newcoords}{adataframe or matrix of coordinates of the new data}

\item{type}{the type of response see \code{\link[glmnet:predict.glmnet]{glmnet::predict.glmnet()}}}
}
\value{
a vector of predicted values
}
\description{
Predict method for GWL fits
}
