% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gwlest.R
\name{print.gwlest}
\alias{print.gwlest}
\title{Printing GWL objects}
\usage{
\method{print}{gwlest}(x, ...)
}
\arguments{
\item{x}{an object of class \code{gwlest}}

\item{...}{ellipsis for S3 method compatibility}
}
\value{
this function print key elements of a \code{gwlest} object
}
\description{
Printing GWL objects
}
