% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gwlfit.R
\name{print.gwlfit}
\alias{print.gwlfit}
\title{Printing GWL objects}
\usage{
\method{print}{gwlfit}(x, ...)
}
\arguments{
\item{x}{a gwlfit object}

\item{...}{ellipsis for S3 method compatibility}
}
\value{
this function print key elements of a \code{gwlfit} object
}
\description{
Printing GWL objects
}
