% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWlasso-package.R
\docType{package}
\name{GWlasso-package}
\alias{GWlasso}
\alias{GWlasso-package}
\title{GWlasso: Geographically Weighted Lasso}
\description{
Performs geographically weighted Lasso regressions. Find optimal bandwidth, fit a geographically weighted lasso or ridge regression, and make predictions. These methods are specially well suited for ecological inferences. Bandwidth selection algorithm is from A. Comber and P. Harris (2018) \doi{10.1007/s10109-018-0280-7}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nibortolum/GWlasso}
  \item \url{https://nibortolum.github.io/GWlasso/}
  \item Report bugs at \url{https://github.com/nibortolum/GWlasso/issues}
}

}
\author{
\strong{Maintainer}: Matthieu Mulot \email{matthieu.mulot@gmail.com} (\href{https://orcid.org/0000-0002-8039-5078}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Sophie Erb (\href{https://orcid.org/0000-0002-0700-283X}{ORCID})
}

}
\keyword{internal}
