% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gwlfit.R
\name{plot.gwlfit}
\alias{plot.gwlfit}
\title{Plot method for gwlfit object}
\usage{
\method{plot}{gwlfit}(x, ...)
}
\arguments{
\item{x}{a gwlfit object returnes by \code{\link[=gwl_fit]{gwl_fit()}}}

\item{...}{ellipsis for S3 method compatibility}
}
\value{
a ggplot
}
\description{
Plot method for gwlfit object
}
\examples{

predictors <- matrix(data = rnorm(20000), 200,100)
y_value <- sample(1:1000, 200)
coords <- data.frame("Lat" = rnorm(200), "Long" = rnorm(200))
distance_matrix <- compute_distance_matrix(coords)

my.gwl.fit <- gwl_fit(bw = 20,
x.var = predictors, 
y.var = y_value,
kernel = "bisquare",
dist.mat = distance_matrix, 
alpha = 1, 
adaptive = TRUE, 
progress = TRUE,
nfolds = 5)

plot(my.gwl.fit)

}
