% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gwlest.R
\name{print.gwlest}
\alias{print.gwlest}
\title{Printing GWL objects}
\usage{
\method{print}{gwlest}(x, ...)
}
\arguments{
\item{x}{an object of class \code{gwlest}}

\item{...}{ellipsis for S3 method compatibility}
}
\value{
this function print key elements of a \code{gwlest} object
}
\description{
Printing GWL objects
}
\examples{

#' predictors <- matrix(data = rnorm(20000), 200,100)
y_value <- sample(1:1000, 200)
coords <- data.frame("Lat" = rnorm(200), "Long" = rnorm(200))
distance_matrix <- compute_distance_matrix(coords)

\dontrun{
  myst.est <- gwl_bw_estimation(x.var = predictors, 
                                y.var = y_value,
                                dist.mat = distance_matrix,
                                adaptive = TRUE,
                                adptbwd.thresh = 0.1,
                                kernel = "bisquare",
                                alpha = 1,
                                progress = TRUE,
                                n=10,
                                nfolds = 5)
  
  
  myst.est
  
}

}
