% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gwlfit.R
\name{print.gwlfit}
\alias{print.gwlfit}
\title{Printing GWL objects}
\usage{
\method{print}{gwlfit}(x, ...)
}
\arguments{
\item{x}{a gwlfit object}

\item{...}{ellipsis for S3 method compatibility}
}
\value{
this function print key elements of a \code{gwlfit} object
}
\description{
Printing GWL objects
}
\examples{
predictors <- matrix(data = rnorm(20000), 200,100)
y_value <- sample(1:1000, 200)
coords <- data.frame("Lat" = rnorm(200), "Long" = rnorm(200))
distance_matrix <- compute_distance_matrix(coords)

my.gwl.fit <- gwl_fit(bw = 20,
x.var = predictors, 
y.var = y_value,
kernel = "bisquare",
dist.mat = distance_matrix, 
alpha = 1, 
adaptive = TRUE,
progress = TRUE,
nfolds = 5)

my.gwl.fit

}
